fix: enable PDK template generation for github-actions files
Only `.sync.yml` is manually modified. Other files are generated by PDK
based on the modification of `.sync.yml`.

Enable the gitlab-action generation, which also contains execution of
the litmus acceptance test.

RedHat and SLES causes errors on provisioning which seems unrelated to
this repo so disable them for now.

Signed-off-by: Claudius Heine <ch@denx.de>