fix(metadata.json): replace 'lamawithonel' with 'siemens'
Since this projects now has a new home in the `siemens` github group,
change the urls and references to 'siemens'.

The `author` field does not support a list of authors, so it has to be
replaced to reflect the current authors of this module. `siemens` is
used here as well, because it is expected that multiple people will
maintain this project in the future.

Signed-off-by: Claudius Heine <ch@denx.de>